[9 12 30 31 36 39] Added getUsersInGroupWithPoints query and changed points value to Float (#31)
* Added getUsersInGroupWithPoints query and changed points value to Floar

* Added aggregate and grouping by category

* Added getStudentPoints function that fetches data tu student's points table

* Added editionId to getStudentPoints

* Added getSumOfPointsForStudentByCategory

* Too much to count, sry

* Cleaned up

* Small bugfix

+++ backend/src/main/kotlin/backend/points/Points.kt
@@ -22,7 +22,7 @@ class Points(
     var teacher: Users,
 
     @Column(name = "value", nullable = false)
-    var value: Long,
+    var value: Float,
 
     @ManyToOne
     @JoinColumn(name = "subcategory_id", referencedColumnName = "subcategory_id")
@@ -35,7 +35,7 @@ class Points(
     constructor() : this(
         student = Users(),
         teacher = Users(),
-        value = 0,
+        value = 0f,
         subcategory = Subcategories(),
         label = ""
     )
@@ -45,6 +45,6 @@ class Points(
 data class PointsInput(
     val student: Long,
     val teacher: Long,
-    val value: Long,
+    val value: Float,
     val subcategory: Long
 )
